// packages/runtime-core/src/enums.ts
export const enum LifecycleHooks {
	BEFORE_CREATE = 'bc',
	CREATED = 'c',
	BEFORE_MOUNT = 'bm',
	MOUNTED = 'm',
	BEFORE_UPDATE = 'bu',
	UPDATED = 'u',
	BEFORE_UNMOUNT = 'bum',
	UNMOUNTED = 'um',
	DEACTIVATED = 'da',
	ACTIVATED = 'a',
	RENDER_TRIGGERED = 'rtg',
	RENDER_TRACKED = 'rtc',
	ERROR_CAPTURED = 'ec',
	SERVER_PREFETCH = 'sp'
}